stream_producer/requirements.txt
kafka-python==2.0.2
Pillow==8.1.2
opencv-python-headless==4.7.0.72